+++ src/card/cards/activation_timing.rs
@@ -6,3 +6,6 @@ pub const DISRUPTING_SCEPTER: CardDefinitionId = CardDefinitionId(1457);
 pub const DWARVEN_WEAPONSMITH: CardDefinitionId = CardDefinitionId(1458);
 pub const SVYELUNITE_PRIEST: CardDefinitionId = CardDefinitionId(1459);
 pub const GWENDLYN_DI_CORCI: CardDefinitionId = CardDefinitionId(1460);
+pub const GATE_TO_PHYREXIA: CardDefinitionId = CardDefinitionId(1461);
+pub const FIRE_DRAKE: CardDefinitionId = CardDefinitionId(1462);
+pub const DARKTHICKET_WOLF: CardDefinitionId = CardDefinitionId(1463);